Remove basic array implementation to use strategies also in bytecode interpreter
